move Pages/FileTree to /Tree
