+++ src/components/Tree/Icons.tsx
@@ -1,0 +1,57 @@
+import React from "react"
+import styled, { css } from "styled-components"
+import theme from "styled-theming"
+import { Colors } from "src/consts/Colors"
+
+const fill = theme("mode", {
+  light: Colors.BLACK,
+  dark: Colors.SILVER,
+})
+
+const hoverFill = theme("mode", {
+  light: Colors.GREY_LIGHTER,
+  dark: Colors.GREY_DARKER,
+})
+
+const Svg = styled.svg`
+  fill: ${fill};
+  /* transition: fill 200ms ease-in-out; */
+  :hover {
+    fill: ${hoverFill};
+  }
+`
+
+const MinusSquareO = (props) => (
+  <Svg {...props} viewBox="64 -65 897 897">
+    <g>
+      <path
+        d="M888 760v0v0v-753v0h-752v0v753v0h752zM888 832h-752q-30 0 -51 -21t-21 -51v-753q0 -29 21 -50.5t51 -21.5h753q29 0 50.5 21.5t21.5 50.5v753q0 30 -21.5 51t-51.5 21v0zM732 347h-442q-14 0 -25 10.5t-11 25.5v0q0 15 11 25.5t25 10.5h442q14 0 25 -10.5t11 -25.5v0
+  q0 -15 -11 -25.5t-25 -10.5z"
+      />
+    </g>
+  </Svg>
+)
+
+const PlusSquareO = (props) => (
+  <Svg {...props} viewBox="64 -65 897 897">
+    <g>
+      <path
+        d="M888 760v0v0v-753v0h-752v0v753v0h752zM888 832h-752q-30 0 -51 -21t-21 -51v-753q0 -29 21 -50.5t51 -21.5h753q29 0 50.5 21.5t21.5 50.5v753q0 30 -21.5 51t-51.5 21v0zM732 420h-184v183q0 15 -10.5 25.5t-25.5 10.5v0q-14 0 -25 -10.5t-11 -25.5v-183h-184
+  q-15 0 -25.5 -11t-10.5 -25v0q0 -15 10.5 -25.5t25.5 -10.5h184v-183q0 -15 11 -25.5t25 -10.5v0q15 0 25.5 10.5t10.5 25.5v183h184q15 0 25.5 10.5t10.5 25.5v0q0 14 -10.5 25t-25.5 11z"
+      />
+    </g>
+  </Svg>
+)
+
+const CloseSquareO = (props) => (
+  <Svg {...props} viewBox="64 -65 897 897">
+    <g>
+      <path
+        d="M717.5 589.5q-10.5 10.5 -25.5 10.5t-26 -10l-154 -155l-154 155q-11 10 -26 10t-25.5 -10.5t-10.5 -25.5t11 -25l154 -155l-154 -155q-11 -10 -11 -25t10.5 -25.5t25.5 -10.5t26 10l154 155l154 -155q11 -10 26 -10t25.5 10.5t10.5 25t-11 25.5l-154 155l154 155
+  q11 10 11 25t-10.5 25.5zM888 760v0v0v-753v0h-752v0v753v0h752zM888 832h-752q-30 0 -51 -21t-21 -51v-753q0 -29 21 -50.5t51 -21.5h753q29 0 50.5 21.5t21.5 50.5v753q0 30 -21.5 51t-51.5 21v0z"
+      />
+    </g>
+  </Svg>
+)
+
+export const Icons = { PlusSquareO, MinusSquareO, CloseSquareO }
